requirements.txt
streamlit
langchain
langchain-community
openai
faiss-cpu
python-dotenv
watchdog
tiktoken
unstructured
python-magic-bin==0.4.14
python-magic
pinecone-client